micro80: config.h type change

+++ Kernel/platform-micro80/config.h
@@ -21,7 +21,7 @@
 #define PROC_SIZE   60	  /* Memory needed per process */
 
 #define SWAPDEV     (swap_dev)	/* A variable for dynamic, or a device major/minor */
-extern unsigned int swap_dev;
+extern uint16_t swap_dev;
 #define SWAP_SIZE   0x79 	/* 60.5K in blocks (prog + udata) */
 #define SWAPBASE    0x1000	/* start at the base of user mem */
 #define SWAPTOP	    0x10000UL	/* Swap out program */
